soc/amd: use expected function signature for soc_acpi_write_tables
A pointer to soc_acpi_write_tables gets assigned to the
write_acpi_tables element of the device_operations struct, so make sure
that the function has the expected function signature which in this case
means using unsigned long as type for both the 'current' parameter and
the return value.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/80218
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/amd/mendocino/agesa_acpi.c
@@ -9,8 +9,8 @@
 #include <soc/acpi.h>
 #include <types.h>
 
-uintptr_t soc_acpi_write_tables(const struct device *device, uintptr_t current,
-				acpi_rsdp_t *rsdp)
+unsigned long soc_acpi_write_tables(const struct device *device, unsigned long current,
+				    acpi_rsdp_t *rsdp)
 {
 	/* TODO: look into adding CRAT */
